Initial reorganization
- Reorganizes VIPS reading into `sf.slide.readers.vips`
