+++ slideflow/slide/__init__.py
@@ -1,8 +1,6 @@
 '''This module includes tools to convolutionally section whole slide images
 into tiles. These tessellated tiles can be exported as PNG or JPG as raw
-images or stored in the binary format TFRecords, with or without augmentation.
-
-Requires: libvips (https://libvips.github.io/libvips/).'''
+images or stored in the binary format TFRecords, with or without augmentation.'''
 
 from __future__ import absolute_import, division, print_function
 
@@ -18,7 +16,8 @@ import warnings
 from functools import partial
 from os.path import exists, join
 from types import SimpleNamespace
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import (Any, Callable, Dict, List, Optional, Tuple, Union,
+                    TYPE_CHECKING)
 
 import cv2
 import numpy as np
@@ -30,62 +29,28 @@ import skimage
 import skimage.filters
 import slideflow as sf
 import slideflow.slide.qc
-from PIL import Image, ImageDraw, UnidentifiedImageError
+from PIL import Image, ImageDraw
+from rich.progress import Progress
 from skimage import img_as_ubyte
 from slideflow import errors
-from slideflow.slide.report import ExtractionPDF  # noqa F401
-from slideflow.slide.report import ExtractionReport, SlideReport
 from slideflow.util import SUPPORTED_FORMATS  # noqa F401
 from slideflow.util import log, path_to_name  # noqa F401
-from rich.progress import Progress
 
-try:
-    import pyvips as vips
-except (ModuleNotFoundError, OSError) as e:
-    log.error("Unable to load vips; slide processing will be unavailable. "
-              f"Error raised: {e}")
+from .report import ExtractionPDF  # noqa F401
+from .report import ExtractionReport, SlideReport
+from .utils import *
+from .readers.vips import (vips2numpy, vips_resize, vips_thumbnail,
+                           _VIPSReader, _JPGVIPSReader)
+
+
+if TYPE_CHECKING:
+    try:
+        import pyvips as vips
+    except Exception:
+        pass
 
 warnings.simplefilter('ignore', Image.DecompressionBombWarning)
 Image.MAX_IMAGE_PIXELS = 100000000000
-DEFAULT_JPG_MPP = 1
-OPS_LEVEL_COUNT = 'openslide.level-count'
-OPS_MPP_X = 'openslide.mpp-x'
-OPS_VENDOR = 'openslide.vendor'
-TIF_EXIF_KEY_MPP = 65326
-OPS_WIDTH = 'width'
-OPS_HEIGHT = 'height'
-DEFAULT_WHITESPACE_THRESHOLD = 230
-DEFAULT_WHITESPACE_FRACTION = 1.0
-DEFAULT_GRAYSPACE_THRESHOLD = 0.05
-DEFAULT_GRAYSPACE_FRACTION = 0.6
-FORCE_CALCULATE_WHITESPACE = -1
-FORCE_CALCULATE_GRAYSPACE = -1
-
-
-def OPS_LEVEL_HEIGHT(level: int) -> str:
-    return f'openslide.level[{level}].height'
-
-
-def OPS_LEVEL_WIDTH(level: int) -> str:
-    return f'openslide.level[{level}].width'
-
-
-def OPS_LEVEL_DOWNSAMPLE(level: int) -> str:
-    return f'openslide.level[{level}].downsample'
-
-
-VIPS_FORMAT_TO_DTYPE = {
-    'uchar': np.uint8,
-    'char': np.int8,
-    'ushort': np.uint16,
-    'short': np.int16,
-    'uint': np.uint32,
-    'int': np.int32,
-    'float': np.float32,
-    'double': np.float64,
-    'complex': np.complex64,
-    'dpcomplex': np.complex128,
-}
 
 
 def _update_kw_with_defaults(kwargs) -> Dict:
@@ -336,40 +301,6 @@ def _wsi_extraction_worker(
     return return_dict
 
 
-def vips2numpy(vi: "vips.Image") -> np.ndarray:
-    '''Converts a VIPS image into a numpy array'''
-    return np.ndarray(buffer=vi.write_to_memory(),
-                      dtype=VIPS_FORMAT_TO_DTYPE[vi.format],
-                      shape=[vi.height, vi.width, vi.bands])
-
-
-def vips_resize(
-    img: np.ndarray,
-    crop_width: int,
-    target_px: int
-) -> np.ndarray:
-    """Resizes and crops an image using libvips.resize()
-
-    Args:
-        img (np.ndarray): Image.
-        crop_width (int): Height/width of image crop (before resize).
-        target_px (int): Target size of final image after resizing.
-
-    Returns:
-        np.ndarray: Resized image.
-    """
-    img_data = np.ascontiguousarray(img).data
-    vips_image = vips.Image.new_from_memory(
-        img_data,
-        crop_width,
-        crop_width,
-        bands=3,
-        format="uchar"
-    )
-    vips_image = vips_image.resize(target_px/crop_width)
-    return vips2numpy(vips_image)
-
-
 def log_extraction_params(**kwargs) -> None:
     '''Logs tile extraction parameters.'''
 
@@ -403,285 +334,6 @@ def log_extraction_params(**kwargs) -> None:
         log.debug(f'Grayspace defined as HSV avg < {gs_t} {excl}')
 
 
-class _VIPSWrapper:
-
-    def __init__(
-        self,
-        path: str,
-        mpp: Optional[float] = None,
-        cache_kw: Optional[Dict[str, Any]] = None
-    ) -> None:
-        '''Wrapper for VIPS to preserve openslide-like functions.'''
-        self.path = path
-        self.cache_kw = cache_kw if cache_kw else {}
-        self.loaded_downsample_levels = {}  # type: Dict[int, "vips.Image"]
-        loaded_image = self.load_downsample_level(0)
-
-        # Load image properties
-        self.properties = {}
-        for field in loaded_image.get_fields():
-            self.properties.update({field: loaded_image.get(field)})
-        self.dimensions = (
-            int(self.properties[OPS_WIDTH]),
-            int(self.properties[OPS_HEIGHT])
-        )
-        # If Openslide MPP is not available, try reading from metadata
-        if mpp is not None:
-            log.debug(f"Setting MPP to {mpp}")
-            self.properties[OPS_MPP_X] = mpp
-        elif OPS_MPP_X not in self.properties.keys():
-            log.debug(
-                "Microns-Per-Pixel (MPP) not found, Searching EXIF"
-            )
-            try:
-                with Image.open(path) as img:
-                    if TIF_EXIF_KEY_MPP in img.tag.keys():
-                        _mpp = img.tag[TIF_EXIF_KEY_MPP][0]
-                        log.debug(
-                            f"Using MPP {_mpp} per EXIF {TIF_EXIF_KEY_MPP}"
-                        )
-                        self.properties[OPS_MPP_X] = _mpp
-                    elif (sf.util.path_to_ext(path).lower() == 'svs'
-                          and 'image-description' in loaded_image.get_fields()):
-                          img_des = loaded_image.get('image-description')
-                          _mpp = re.findall(r'(?<=MPP\s\=\s)0\.\d+', img_des)
-                          if _mpp is not None:
-                            log.debug(
-                                f"Using MPP {_mpp} from 'image-description' for SCN"
-                                "-converted SVS format"
-                            )
-                            self.properties[OPS_MPP_X] = _mpp[0]
-                    elif (sf.util.path_to_ext(path).lower() in ('tif', 'tiff')
-                          and 'xres' in loaded_image.get_fields()):
-                        xres = loaded_image.get('xres')  # 4000.0
-                        if (xres == 4000.0
-                           and loaded_image.get('resolution-unit') == 'cm'):
-                            # xres = xres # though resolution from tiffinfo
-                            # says 40000 pixels/cm, for some reason the xres
-                            # val is 4000.0, so multipley by 10.
-                            # Convert from pixels/cm to cm/pixels, then convert
-                            # to microns by multiplying by 1000
-                            mpp_x = (1/xres) * 1000
-                            self.properties[OPS_MPP_X] = str(mpp_x)
-                            log.debug(
-                                f"Using MPP {mpp_x} per TIFF 'xres' field"
-                                f" {loaded_image.get('xres')} and "
-                                f"{loaded_image.get('resolution-unit')}"
-                            )
-                    else:
-                        name = path_to_name(path)
-                        log.warning(
-                            f"Missing Microns-Per-Pixel (MPP) for {name}"
-                        )
-            except AttributeError:
-                mpp = DEFAULT_JPG_MPP
-                log.debug(f"Could not detect microns-per-pixel; using default {mpp}")
-                self.properties[OPS_MPP_X] = mpp
-            except UnidentifiedImageError:
-                log.error(
-                    f"PIL error; unable to read slide {path_to_name(path)}."
-                )
-
-        if OPS_LEVEL_COUNT in self.properties:
-            self.level_count = int(self.properties[OPS_LEVEL_COUNT])
-            # Calculate level metadata
-            self.levels = []   # type: List[Dict[str, Any]]
-            for lev in range(self.level_count):
-                width = int(loaded_image.get(OPS_LEVEL_WIDTH(lev)))
-                height = int(loaded_image.get(OPS_LEVEL_HEIGHT(lev)))
-                downsample = float(loaded_image.get(OPS_LEVEL_DOWNSAMPLE(lev)))
-                self.levels += [{
-                    'dimensions': (width, height),
-                    'width': width,
-                    'height': height,
-                    'downsample': downsample
-                }]
-        else:
-            self.level_count = 1
-            self.levels = [{
-                    'dimensions': self.dimensions,
-                    'width': self.dimensions[0],
-                    'height': self.dimensions[1],
-                    'downsample': 1
-                }]
-        self.level_downsamples = [lev['downsample'] for lev in self.levels]
-        self.level_dimensions = [lev['dimensions'] for lev in self.levels]
-
-    def best_level_for_downsample(
-        self,
-        downsample: float,
-    ) -> int:
-        '''Return lowest magnification level with a downsample level lower than
-        the given target.
-
-        Args:
-            downsample (float): Ratio of target resolution to resolution
-                at the highest magnification level. The downsample level of the
-                highest magnification layer is equal to 1.
-            levels (list(int), optional): Valid levels to search. Defaults to
-                None (search all levels).
-
-        Returns:
-            int:    Optimal downsample level.'''
-        max_downsample = 0
-        for d in self.level_downsamples:
-            if d < downsample:
-                max_downsample = d
-        try:
-            max_level = self.level_downsamples.index(max_downsample)
-        except Exception:
-            log.debug(f"Error attempting to read level {max_downsample}")
-            return 0
-        return max_level
-
-    def load_downsample_level(self, level: int) -> "vips.Image":
-        downsampled_image = vips.Image.new_from_file(
-            self.path,
-            level=level,
-            fail=True,
-            access=vips.enums.Access.RANDOM
-        )
-        if self.cache_kw:
-            downsampled_image = downsampled_image.tilecache(**self.cache_kw)
-        self.loaded_downsample_levels.update({
-            level: downsampled_image
-        })
-        return downsampled_image
-
-    def get_downsampled_image(self, level: int) -> "vips.Image":
-        '''Returns a VIPS image of a given downsample.'''
-        if level in range(len(self.levels)):
-            if level in self.loaded_downsample_levels:
-                return self.loaded_downsample_levels[level]
-            else:
-                return self.load_downsample_level(level)
-        else:
-            return False
-
-    def read_region(
-        self,
-        base_level_dim: Tuple[int, int],
-        downsample_level: int,
-        extract_size: Tuple[int, int]
-    ) -> "vips.Image":
-        """Extracts a region from the image at the given downsample level.
-
-        Args:
-            base_level_dim (Tuple[int, int]): Top-left location of the region
-                to extract, using downsample layer coordinates (x, y)
-            downsample_level (int): Downsample level to read.
-            extract_size (Tuple[int, int]): Size of the region to read
-                (width, height) using base layer coordinates.
-
-        Returns:
-            vips.Image: VIPS image.
-        """
-        base_level_x, base_level_y = base_level_dim
-        extract_width, extract_height = extract_size
-        downsample_factor = self.level_downsamples[downsample_level]
-        downsample_x = int(base_level_x / downsample_factor)
-        downsample_y = int(base_level_y / downsample_factor)
-        image = self.get_downsampled_image(downsample_level)
-        region = image.crop(
-            downsample_x,
-            downsample_y,
-            extract_width,
-            extract_height
-        )
-        return region
-
-    def read_from_pyramid(
-        self,
-        top_left: Tuple[int, int],
-        window_size: Tuple[int, int],
-        target_size: Optional[Tuple[int, int]] = None,
-        target_downsample: Optional[float] = None,
-    ) -> "vips.Image":
-        """Reads a region from the image. Performance is accelerated by
-        pyramid downsample layers, if available.
-
-        Args:
-            top_left (Tuple[int, int]): Top-left location of the region to
-                extract, using base layer coordinates (x, y).
-            window_size (Tuple[int, int]): Size of the region to read (width,
-                height) using base layer coordinates.
-            target_size (Tuple[int, int]): Resize the region to this target
-                size (width, height).
-
-        Returns:
-            vips.Image: VIPS image. Dimensions will equal target_size unless
-            the window includes an area of the image which is out of bounds.
-            In this case, the returned image will be cropped.
-        """
-        if target_size is None and target_downsample is None:
-            raise ValueError("Must supply either target_size or "
-                             "target_downsample to read_from_pyramid()")
-        if target_downsample is None:
-            target_downsample = window_size[0] / target_size[0]
-
-        ds_level = self.best_level_for_downsample(target_downsample)
-        image = self.get_downsampled_image(ds_level)
-        resize_factor = self.level_downsamples[ds_level] / target_downsample
-        image = image.resize(resize_factor)
-
-        if target_size is not None:
-            return image.crop(
-                int(top_left[0] / target_downsample),
-                int(top_left[1] / target_downsample),
-                min(target_size[0], image.width),
-                min(target_size[1], image.height)
-            )
-        else:
-            return image
-
-class _JPGslideToVIPS(_VIPSWrapper):
-    '''Wrapper for JPG files, which do not possess separate levels, to
-    preserve openslide-like functions.'''
-
-    def __init__(self, path: str, mpp: Optional[float] = None, cache_kw = None) -> None:
-        self.path = path
-        self.full_image = vips.Image.new_from_file(path)
-        self.cache_kw = cache_kw if cache_kw else {}
-        if not self.full_image.hasalpha():
-            self.full_image = self.full_image.addalpha()
-        self.properties = {}
-        for field in self.full_image.get_fields():
-            self.properties.update({field: self.full_image.get(field)})
-        width = int(self.properties[OPS_WIDTH])
-        height = int(self.properties[OPS_HEIGHT])
-        self.dimensions = (width, height)
-        self.level_count = 1
-        self.loaded_downsample_levels = {
-            0: self.full_image
-        }
-        # Calculate level metadata
-        self.levels = [{
-            'dimensions': (width, height),
-            'width': width,
-            'height': height,
-            'downsample': 1,
-        }]
-        self.level_downsamples = [1]
-        self.level_dimensions = [(width, height)]
-
-        # MPP data
-        if mpp is not None:
-            log.debug(f"Setting MPP to {mpp}")
-            self.properties[OPS_MPP_X] = mpp
-        else:
-            try:
-                with Image.open(path) as img:
-                    exif_data = img.getexif()
-                    if TIF_EXIF_KEY_MPP in exif_data.keys():
-                        _mpp = exif_data[TIF_EXIF_KEY_MPP]
-                        log.debug(f"Using MPP {_mpp} per EXIF field {TIF_EXIF_KEY_MPP}")
-                        self.properties[OPS_MPP_X] = _mpp
-                    else:
-                        raise AttributeError
-            except AttributeError:
-                mpp = DEFAULT_JPG_MPP
-                log.debug(f"Could not detect microns-per-pixel; using default {mpp}")
-                self.properties[OPS_MPP_X] = mpp
 class ROI:
     '''Object container for ROI annotations.'''
 
@@ -825,11 +477,11 @@ class _BaseLoader:
         self.__dict__.update(state)
 
     @property
-    def _vips_wrapper(self) -> Any:
+    def _slide_reader(self) -> Any:
         if self.filetype.lower() in ('jpg', 'jpeg'):
-            return _JPGslideToVIPS
+            return _JPGVIPSReader
         else:
-            return _VIPSWrapper
+            return _VIPSReader
 
     @property
     def dimensions(self) -> Tuple[int, int]:
@@ -847,17 +499,17 @@ class _BaseLoader:
             return None
 
     @property
-    def slide(self) -> _VIPSWrapper:
+    def slide(self) -> Union[_JPGVIPSReader, _VIPSReader]:
         if self.__slide is not None:
             return self.__slide
 
         try:
-            self.__slide = self._vips_wrapper(
+            self.__slide = self._slide_reader(
                 self.path,
                 self._mpp_override,
                 self._vips_cache_kw)
             return self.__slide  # type: ignore
-        except vips.error.Error as e:
+        except Exception as e:
             raise errors.SlideLoadError(
                 f"Error loading slide {self.shortname}: {e}"
             )
@@ -989,7 +641,12 @@ class _BaseLoader:
         self.estimated_num_tiles = int(self.grid.sum())
         return Image.fromarray(img_as_ubyte(self.qc_mask))
 
-    def square_thumb(self, width: int = 512) -> Image.Image:
+    def square_thumb(
+        self,
+        width: int = 512,
+        use_associated_image: bool = True,
+        **kwargs
+    ) -> Image.Image:
         '''Returns a square thumbnail of the slide, with black bar borders.
 
         Args:
@@ -998,23 +655,12 @@ class _BaseLoader:
         Returns:
             PIL image
         '''
-        # Get thumbnail image and dimensions via fastest method available
-        associated_images = self.slide.properties['slide-associated-images']
-        if ('slide-associated-images' in self.slide.properties
-           and 'thumbnail' in associated_images):
-            vips_thumb = vips.Image.openslideload(
-                self.slide.path,
-                associated='thumbnail'
-            )
-        else:
-            level = max(0, self.slide.level_count-2)
-            vips_thumb = self.slide.get_downsampled_image(level)
-
-        height = int(width / (vips_thumb.width / vips_thumb.height))
-        np_thumb = vips2numpy(vips_thumb)
-        thumb = Image.fromarray(np_thumb).resize((width, height))
-
-        # Standardize to square with black borders as needed
+        thumb = self.thumb(
+            width=width,
+            use_associated_image=use_associated_image,
+            **kwargs)
+        height = int(width / (thumb.width / thumb.height))
+        thumb = thumb.resize((width, height))
         square_thumb = Image.new("RGB", (width, width))
         square_thumb.paste(thumb, (0, int((width-height)/2)))
         return square_thumb
@@ -1024,9 +670,9 @@ class _BaseLoader:
         mpp: Optional[float] = None,
         width: Optional[int] = None,
         coords: Optional[List[int]] = None,
-        rois: bool = False,
-        linewidth: int = 2,
-        color: str = 'black'
+        rect_linewidth: int = 2,
+        rect_color: str = 'black',
+        use_associated_image: bool = False
     ) -> Image.Image:
         '''Returns PIL thumbnail of the slide.
 
@@ -1038,6 +684,8 @@ class _BaseLoader:
             coords (list(int), optional): List of tile extraction coordinates
                 to show as rectangles on the thumbnail, in [(x_center,
                 y_center), ...] format. Defaults to None.
+            use_associated_image (bool): Use the associated thumbnail image
+                in the slide, rather than reading from a pyramid layer.
 
         Returns:
             PIL image
@@ -1063,23 +711,15 @@ class _BaseLoader:
         # Calculate appropriate height
         height = int((self.mpp * self.dimensions[1]) / mpp)
 
-        # Get thumb via libvips & convert PIL Image
-        if self.vendor and self.vendor == 'leica':
-            # The libvips thumbnail function does not work appropriately
-            # with Leica SCN images, so a downsample level must be
-            # manually specified.
-            thumbnail = vips.Image.new_from_file(
-                self.path,
-                fail=True,
-                access=vips.enums.Access.RANDOM,
-                level=self.slide.level_count-1
-            )
+        if use_associated_image:
+            thumb_kw = dict(associated='thumbnail')
         else:
-            thumbnail = vips.Image.thumbnail(self.path, width)
-        try:
-            np_thumb = vips2numpy(thumbnail)
-        except vips.error.Error as e:
-            raise errors.SlideLoadError(f"Error loading slide thumbnail: {e}")
+            thumb_kw = dict(level=self.slide.level_count-1, width=width)
+
+        np_thumb = vips_thumbnail(
+            self.path,
+            leica_scn=(self.vendor and self.vendor == 'leica'),
+            **thumb_kw)
         image = Image.fromarray(np_thumb).resize((width, height))
 
         if coords:
@@ -1089,7 +729,7 @@ class _BaseLoader:
             for (x, y) in coords:  # type: ignore
                 x, y = x * ratio * self.roi_scale, y * ratio * self.roi_scale  # type: ignore
                 coords = (x-wh, y-wh, x+wh, y+wh)  # type: ignore
-                draw.rectangle(coords, outline='black', width=2)
+                draw.rectangle(coords, outline=rect_color, width=rect_linewidth)
             return image
         else:
             return image
@@ -1552,7 +1192,7 @@ class WSI(_BaseLoader):
             (x, y, grid_x, grid_y),
             SimpleNamespace(
                 full_extract_px=self.full_extract_px,
-                vips_wrapper=self._vips_wrapper,
+                vips_wrapper=self._slide_reader,
                 mpp_override=self._mpp_override,
                 vips_cache=self._vips_cache_kw,
                 roi_scale=self.roi_scale,
@@ -1829,7 +1469,7 @@ class WSI(_BaseLoader):
 
         w_args = SimpleNamespace(**{
             'full_extract_px': self.full_extract_px,
-            'vips_wrapper': self._vips_wrapper,
+            'vips_wrapper': self._slide_reader,
             'mpp_override': self._mpp_override,
             'vips_cache': self._vips_cache_kw,
             'roi_scale': self.roi_scale,
@@ -1955,7 +1595,8 @@ class WSI(_BaseLoader):
         coords: Optional[List[int]] = None,
         rois: bool = False,
         linewidth: int = 2,
-        color: str = 'black'
+        color: str = 'black',
+        use_associated_image: bool = False
     ) -> Image.Image:
         """Returns PIL Image of thumbnail with ROI overlay.
 
@@ -1989,7 +1630,11 @@ class WSI(_BaseLoader):
             else:
                 roi_scale = self.dimensions[0] / width  # type: ignore
 
-        thumb = super().thumb(mpp=mpp, width=width, coords=coords)
+        thumb = super().thumb(
+            mpp=mpp,
+            width=width,
+            coords=coords,
+            use_associated_image=use_associated_image)
 
         if rois:
             annPolys = [
