+++ slideflow/slide/utils.py
@@ -1,0 +1,27 @@
+"""Utility functions and constants for slide reading."""
+
+DEFAULT_JPG_MPP = 1
+OPS_LEVEL_COUNT = 'openslide.level-count'
+OPS_MPP_X = 'openslide.mpp-x'
+OPS_VENDOR = 'openslide.vendor'
+TIF_EXIF_KEY_MPP = 65326
+OPS_WIDTH = 'width'
+OPS_HEIGHT = 'height'
+DEFAULT_WHITESPACE_THRESHOLD = 230
+DEFAULT_WHITESPACE_FRACTION = 1.0
+DEFAULT_GRAYSPACE_THRESHOLD = 0.05
+DEFAULT_GRAYSPACE_FRACTION = 0.6
+FORCE_CALCULATE_WHITESPACE = -1
+FORCE_CALCULATE_GRAYSPACE = -1
+
+
+def OPS_LEVEL_HEIGHT(level: int) -> str:
+    return f'openslide.level[{level}].height'
+
+
+def OPS_LEVEL_WIDTH(level: int) -> str:
+    return f'openslide.level[{level}].width'
+
+
+def OPS_LEVEL_DOWNSAMPLE(level: int) -> str:
+    return f'openslide.level[{level}].downsample'
